clip3 param can be used to specify a secondary external clip to deinterlace the OTHER field.
Example: If the video is bff then yadif keeps the bottom field. When we need the top field for a single frame, there is no way to change it through TFM. Unless there is a clip3 ready, with all the top fields deinterlaced. A bit wasteful, tell me if there is a better way.